Minor code improvements

+++ src/commands/allowlist.rs
@@ -27,19 +27,20 @@ async fn allowlist(ctx: &Context, msg: &Message, args: Args) -> CommandResult {
     {
         let guild_id = msg.guild_id.unwrap().0;
         let data = ctx.data.read().await;
-        let mutex = data.get::<Database>().unwrap();
-        let db = mutex.lock().unwrap();
-        let number_of_allowlisted_users = match db.count_allowlisted_users_in_guild(guild_id) {
-            Ok(n) => n,
-            Err(e) => return Err(CommandError::from(e.to_string())),
-        };
-        if number_of_allowlisted_users >= MAXIMUM_NUMBER_OF_ALLOWLISTED_USERS_PER_GUILD {
-            return Err(CommandError::from("Reached maximum number of allowlisted users"));
+        if let Some(mutex) = data.get::<Database>() {
+            let db = mutex.lock().unwrap();
+            let number_of_allowlisted_users = match db.count_allowlisted_users_in_guild(guild_id) {
+                Ok(n) => n,
+                Err(e) => return Err(CommandError::from(e.to_string())),
+            };
+            if number_of_allowlisted_users >= MAXIMUM_NUMBER_OF_ALLOWLISTED_USERS_PER_GUILD {
+                return Err(CommandError::from("Reached maximum number of allowlisted users"));
+            }
+            match db.insert_in_allowlist(msg.guild_id.unwrap().0, user_id) {
+                Ok(_) => (),
+                Err(e) => return Err(CommandError::from(e.to_string())),
+            };
         }
-        match db.insert_in_allowlist(msg.guild_id.unwrap().0, user_id) {
-            Ok(_) => (),
-            Err(e) => return Err(CommandError::from(e.to_string())),
-        };
     }
     log(ctx, msg, format!("Added id={} to allowlist", user_id));
     return Ok(());
@@ -59,12 +60,13 @@ async fn unallowlist(ctx: &Context, msg: &Message, args: Args) -> CommandResult 
     };
     {
         let data = ctx.data.read().await;
-        let mutex = data.get::<Database>().unwrap();
-        let db = mutex.lock().unwrap();
-        match db.remove_from_allowlist(msg.guild_id.unwrap().0, user_id) {
-            Ok(b) => b,
-            Err(e) => return Err(CommandError::from(e.to_string())),
-        };
+        if let Some(mutex) = data.get::<Database>() {
+            let db = mutex.lock().unwrap();
+            match db.remove_from_allowlist(msg.guild_id.unwrap().0, user_id) {
+                Ok(b) => b,
+                Err(e) => return Err(CommandError::from(e.to_string())),
+            };
+        }
     }
     return Ok(());
 }
@@ -84,12 +86,13 @@ async fn is_allowlisted(ctx: &Context, msg: &Message, args: Args) -> CommandResu
     let is_allowlisted: bool;
     {
         let data = ctx.data.read().await;
-        let mutex = data.get::<Database>().unwrap();
-        let db = mutex.lock().unwrap();
-        is_allowlisted = match db.is_allowlisted(msg.guild_id.unwrap().0, user_id) {
-            Ok(b) => b,
-            Err(e) => return Err(CommandError::from(e.to_string())),
-        };
+        if let Some(mutex) = data.get::<Database>() {
+            let db = mutex.lock().unwrap();
+            is_allowlisted = match db.is_allowlisted(msg.guild_id.unwrap().0, user_id) {
+                Ok(b) => b,
+                Err(e) => return Err(CommandError::from(e.to_string())),
+            };
+        }
     }
     msg.reply(ctx, format!("{}", is_allowlisted)).await?;
     return Ok(());
@@ -102,12 +105,13 @@ async fn get_allowlisted_users(ctx: &Context, msg: &Message) -> CommandResult {
     let allowlisted_users: Vec<u64>;
     {
         let data = ctx.data.read().await;
-        let mutex = data.get::<Database>().unwrap();
-        let db = mutex.lock().unwrap();
-        allowlisted_users = match db.get_allowlisted_users_in_guild(msg.guild_id.unwrap().0) {
-            Ok(users) => users,
-            Err(e) => return Err(CommandError::from(e.to_string())),
-        };
+        if let Some(mutex) = data.get::<Database>() {
+            let db = mutex.lock().unwrap();
+            allowlisted_users = match db.get_allowlisted_users_in_guild(msg.guild_id.unwrap().0) {
+                Ok(users) => users,
+                Err(e) => return Err(CommandError::from(e.to_string())),
+            };
+        }
     }
     if allowlisted_users.is_empty() {
         msg.reply(ctx, "There is currently no users in the allowlist for this guild.").await?;
